modularize more unary ops
