+++ tfjs-core/src/gradients/Exp_grad.ts
@@ -14,35 +14,17 @@
  * limitations under the License.
  * =============================================================================
  */
-import {Selu} from '../kernel_names';
+
+import {Exp} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
-import {cast} from '../ops/array_ops';
-import {exp} from '../ops/exp';
-import {greater} from '../ops/greater';
 import {mul} from '../ops/mul';
-import {SELU_SCALE, SELU_SCALEALPHA} from '../ops/selu_util';
-import {scalar} from '../ops/tensor_ops';
-import {where} from '../ops/where';
 import {Tensor} from '../tensor';
 
-export const seluGradConfig: GradConfig = {
-  kernelName: Selu,
-  inputsToSave: ['x'],
+export const expGradConfig: GradConfig = {
+  kernelName: Exp,
+  outputsToSave: [true],
   gradFunc: (dy: Tensor, saved: Tensor[]) => {
-    const [x] = saved;
-    return {
-      x: () => {
-        const mask = greater(x, scalar(0));
-
-        const scaleAlpha = scalar(SELU_SCALEALPHA);
-        const scale = scalar(SELU_SCALE);
-
-        const greaterThanZeroDer = mul(dy, scale);
-        const lessEqualZeroDer =
-            mul(mul(dy, scaleAlpha), exp(cast(x, 'float32')));
-
-        return where(mask, greaterThanZeroDer, lessEqualZeroDer);
-      }
-    };
+    const [y] = saved;
+    return {x: () => mul(dy, y)};
   }
 };
