+++ tfjs-core/src/ops/exp.ts
@@ -1,0 +1,48 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import {ENGINE} from '../engine';
+import {Exp, ExpInputs} from '../kernel_names';
+import {Tensor} from '../tensor';
+import {NamedTensorMap} from '../tensor_types';
+import {convertToTensor} from '../tensor_util_env';
+import {TensorLike} from '../types';
+
+import {op} from './operation';
+
+/**
+ * Computes exponential of the input `tf.Tensor` element-wise. `e ^ x`
+ *
+ * ```js
+ * const x = tf.tensor1d([1, 2, -3]);
+ *
+ * x.exp().print();  // or tf.exp(x)
+ * ```
+ * @param x The input tensor.
+ */
+/** @doc {heading: 'Operations', subheading: 'Basic math'} */
+function exp_<T extends Tensor>(x: T|TensorLike): T {
+  const $x = convertToTensor(x, 'x', 'exp');
+
+  const inputs: ExpInputs = {x: $x};
+  return ENGINE.runKernelFunc((backend, save) => {
+    const res = backend.exp($x);
+    save([res]);
+    return res;
+  }, inputs as {} as NamedTensorMap, null /* grad */, Exp);
+}
+export const exp = op({exp_});
