+++ tfjs-core/src/ops/log_loss.ts
@@ -22,13 +22,13 @@ import {assertShapesMatch} from '../util';
 
 import {add} from './add';
 import {computeWeightedLoss} from './compute_weighted_loss';
+import {log} from './log';
 import {Reduction} from './loss_ops_utils';
 import {mul} from './mul';
 import {neg} from './neg';
 import {op} from './operation';
 import {sub} from './sub';
 import {scalar} from './tensor_ops';
-import {log} from './unary_ops';
 
 /**
  * Computes the log loss between two tensors.
