+++ tfjs-core/src/ops/ops.ts
@@ -40,6 +40,7 @@ export {batchNorm3d} from './batchnorm3d';
 export {batchNorm4d} from './batchnorm4d';
 export {broadcastTo} from './broadcast_to';
 export {ceil} from './ceil';
+export {clipByValue} from './clip_by_value';
 export {clone} from './clone';
 export {complex} from './complex';
 export {concat} from './concat';
@@ -64,7 +65,10 @@ export {divNoNan} from './div_no_nan';
 export {dot} from './dot';
 export {elu} from './elu';
 export {equal} from './equal';
+export {erf} from './erf';
+export {exp} from './exp';
 export {expandDims} from './expand_dims';
+export {expm1} from './expm1';
 export {eye} from './eye';
 export {fill} from './fill';
 export {floor} from './floor';
@@ -77,6 +81,9 @@ export {leakyRelu} from './leaky_relu';
 export {less} from './less';
 export {lessEqual} from './less_equal';
 export {localResponseNormalization} from './local_response_normalization';
+export {log} from './log';
+export {log1p} from './log1p';
+export {logSigmoid} from './log_sigmoid';
 export {logSumExp} from './log_sum_exp';
 export {logicalAnd} from './logical_and';
 export {logicalNot} from './logical_not';
@@ -114,6 +121,7 @@ export {randomGamma} from './random_gamma';
 export {randomNormal} from './random_normal';
 export {randomUniform} from './random_uniform';
 export {real} from './real';
+export {reciprocal} from './reciprocal';
 export {relu} from './relu';
 export {relu6} from './relu6';
 export {reshape} from './reshape';
