+++ tfjs-core/src/register_all_gradients.ts
@@ -32,6 +32,7 @@ import {batchMatMulGradConfig} from './gradients/BatchMatMul_grad';
 import {batchToSpaceNDGradConfig} from './gradients/BatchToSpaceND_grad';
 import {broadcastToGradConfig} from './gradients/BroadcastTo_grad';
 import {ceilGradConfig} from './gradients/Ceil_grad';
+import {clipByValueGradConfig} from './gradients/ClipByValue_grad';
 import {concatGradConfig} from './gradients/Concat_grad';
 import {conv2DGradConfig} from './gradients/Conv2D_grad';
 import {conv2DBackpropInputGradConfig} from './gradients/Conv2DBackpropInput_grad';
@@ -43,12 +44,17 @@ import {depthwiseConv2dNativeGradConfig} from './gradients/DepthwiseConv2dNative
 import {dilation2dGradConfig} from './gradients/Dilation2D_grad';
 import {divGradConfig} from './gradients/Div_grad';
 import {eluGradConfig} from './gradients/Elu_grad';
+import {erfGradConfig} from './gradients/Erf_grad';
+import {expGradConfig} from './gradients/Exp_grad';
+import {expm1GradConfig} from './gradients/Expm1_grad';
 import {floorGradConfig} from './gradients/Floor_grad';
 import {floorDivGradConfig} from './gradients/FloorDiv_grad';
 import {fusedBatchNormGradConfig} from './gradients/FusedBatchNorm_grad';
 import {gatherGradConfig} from './gradients/GatherV2_grad';
 import {greaterEqualGradConfig} from './gradients/GreaterEqual_grad';
 import {identityGradConfig} from './gradients/Identity_grad';
+import {log1pGradConfig} from './gradients/Log1p_grad';
+import {logGradConfig} from './gradients/Log_grad';
 import {lrnGradConfig} from './gradients/LRN_grad';
 import {maxGradConfig} from './gradients/Max_grad';
 import {maximumGradConfig} from './gradients/Maximum_grad';
@@ -63,6 +69,7 @@ import {oneHotGradConfig} from './gradients/OneHot_grad';
 import {padV2GradConfig} from './gradients/PadV2_grad';
 import {powGradConfig} from './gradients/Pow_grad';
 import {preluGradConfig} from './gradients/Prelu_grad';
+import {reciprocalGradConfig} from './gradients/Reciprocal_grad';
 import {relu6GradConfig} from './gradients/Relu6_grad';
 import {reluGradConfig} from './gradients/Relu_grad';
 import {reshapeGradConfig} from './gradients/Reshape_grad';
@@ -110,6 +117,7 @@ const gradConfigs: GradConfig[] = [
   batchToSpaceNDGradConfig,
   broadcastToGradConfig,
   ceilGradConfig,
+  clipByValueGradConfig,
   concatGradConfig,
   conv2DBackpropInputGradConfig,
   conv2DGradConfig,
@@ -121,12 +129,17 @@ const gradConfigs: GradConfig[] = [
   dilation2dGradConfig,
   divGradConfig,
   eluGradConfig,
+  erfGradConfig,
+  expGradConfig,
+  expm1GradConfig,
   floorDivGradConfig,
   floorGradConfig,
   fusedBatchNormGradConfig,
   gatherGradConfig,
   greaterEqualGradConfig,
   identityGradConfig,
+  log1pGradConfig,
+  logGradConfig,
   lrnGradConfig,
   maxGradConfig,
   maxGradConfig,
@@ -144,6 +157,7 @@ const gradConfigs: GradConfig[] = [
   padV2GradConfig,
   powGradConfig,
   preluGradConfig,
+  reciprocalGradConfig,
   relu6GradConfig,
   reluGradConfig,
   reshapeGradConfig,
@@ -153,9 +167,9 @@ const gradConfigs: GradConfig[] = [
   selectV2PoolGradConfig,
   seluGradConfig,
   signGradConfig,
-  sliceGradConfig,
   sinGradConfig,
   sinhGradConfig,
+  sliceGradConfig,
   spaceToBatchNDGradConfig,
   spaceToBatchNDGradConfig,
   splitVGradConfig,
