+++ tfjs-core/src/tests.ts
@@ -66,6 +66,7 @@ import './ops/boolean_mask_test';
 import './ops/broadcast_to_test';
 import './ops/broadcast_util_test';
 import './ops/ceil_test';
+import './ops/clip_by_value_test';
 import './ops/clone_test';
 import './ops/compare_ops_test';
 import './ops/complex_ops_test';
@@ -93,7 +94,10 @@ import './ops/dropout_test';
 import './ops/dropout_util_test';
 import './ops/elu_test';
 import './ops/equal_test';
+import './ops/erf_test';
+import './ops/exp_test';
 import './ops/expand_dims_test';
+import './ops/expm1_test';
 import './ops/eye_test';
 import './ops/floor_test';
 import './ops/frame_test';
@@ -112,8 +116,11 @@ import './ops/leaky_relu_test';
 import './ops/less_equal_test';
 import './ops/less_test';
 import './ops/local_response_normalization_test';
+import './ops/log1p_test';
 import './ops/log_loss_test';
+import './ops/log_sigmoid_test';
 import './ops/log_sum_exp_test';
+import './ops/log_test';
 import './ops/logical_and_test';
 import './ops/logical_not_test';
 import './ops/logical_or_test';
@@ -145,6 +152,7 @@ import './ops/rand_test';
 import './ops/random_gamma_test';
 import './ops/random_normal_test';
 import './ops/random_uniform_test';
+import './ops/reciprocal_test';
 import './ops/relu6_test';
 import './ops/relu_test';
 import './ops/resize_bilinear_test';
